Ice cannon and frozen Keen (#29)
* Add ice sprites

* Place Ice cannon in debug level

For easy testing

* Ice chunk and frozen signal

Still need to add animation when ice chunk hits a wall

* Have ice cannon spawn chunks

Initial implementation, should also allow a left firing cannon type to be used (only 1 of them is in the game)

* Handle Frozen signal

* Very basic iced state implementation

Currently Keen is frozen and stays like that hanging in mid air...

* Have ice break into pieces

* Restyle lvl 9 right side

* Add thaw animation and related code

Keen can now defrost, timings may need adjusting

* Recreate area to access secret teleporter

To check physics are correct

* Create left facing ice cannon

* Make timer to be one shot

* Have left facing ice cannon chunks

* Add left ice cannon to tile map

* Add missing ice cannon to level

* More cloned areas for debugging

* Move physics process to later in the stack

We need to handle stat change first then run the physic process. This fixes a flaw where there is a frame shown of the character facing the wrong way for example if the raygun is used in the other direction previously.