Level 16 finish (#32)
* Add background

* Create method to insta-kill vorticon

* Create stone block

* Tweak Keen pogo animation

* Add background and stone block

* Debug stone block and vorticon

* Add a platform layer

* Swap AnimatedSprite for Sprite + AnimationPlayer

* Have raygun shots hit platforms

* Make chain shootable

And cause the block to fall (& not get stuck on walls)

* Test block with vorticon

* Add missing invisible blocks at level end

* Ensure block works

Also remove debug messages

* Add end of level effects

* Godot noise

* Godot noise

* Change ShipPart signal to be less focused

Initial work to update all of the ship part UI in case of death in level

* Change pogo stick status to be an enum

Allow pogo status to be altered if Keen dies before finishing the level if he just gained it.

* Introduce game logic to handle death better

Resets the ship parts to before level started if Keen dies
Resets pogo is just gained in level and dies, or ensures it is kept if Keen finishes the level (could have done same as ship parts logic).